Address sign-mismatched comparisons

+++ src/fswAlgorithms/_GeneralModuleFiles/srukfInterface.cpp
@@ -63,7 +63,7 @@ void SRukfInterface::timeUpdate(double updateTime) {
 
     /*! - For each Sigma point, apply sBar-based error, propagate forward, and scale by Wm just like 0th.
      Note that we perform +/- sigma points simultaneously in loop to save loop values.*/
-    for (size_t i = 1; i < this->state.size() + 1; ++i) {
+    for (long i = 1; i < this->state.size() + 1; ++i) {
         /*! - Adding covariance columns from sigma points*/
         this->sigmaPoints[i] = dynamics.propagate(time, this->state.addVector(this->eta * this->sBar.col(i - 1)), dt);
         /*! - Subtracting covariance columns from sigma points*/
